chore(weave): extract stats.py and mutations.py from calls_query_builder
Extract stats query builders and mutation (UPDATE/DELETE) query builders
into dedicated modules.

- stats.py: build_calls_stats_query and all optimized stats helpers
- mutations.py: build_calls_complete_update/delete queries,
  _format_table_name_with_cluster
