+++ weave/trace_server/calls_query_builder/calls_query_builder.py
@@ -27,7 +27,6 @@ Outstanding Optimizations/Work:
 
 import logging
 import re
-from collections.abc import KeysView
 from typing import Literal, NamedTuple, cast
 
 from pydantic import BaseModel, Field
@@ -82,7 +81,6 @@ from weave.trace_server.calls_query_builder.utils import (
     param_slot,
     safely_format_sql,
 )
-from weave.trace_server.clickhouse_trace_server_settings import LOCAL_TABLE_SUFFIX
 from weave.trace_server.common_interface import SortBy
 from weave.trace_server.interface import query as tsi_query
 from weave.trace_server.orm import (
@@ -1369,365 +1367,3 @@ def process_children_of_eval_ids_to_sql(
         f" AND ({parent_id_conditions})"
         f" AND {table_alias}.id NOT IN {eval_root_ids_param}"
     )
-
-
-######### STATS QUERY HANDLING ##########
-
-
-def build_calls_stats_query(
-    req: tsi.CallsQueryStatsReq,
-    param_builder: ParamBuilder,
-    read_table: ReadTable = ReadTable.CALLS_MERGED,
-) -> tuple[str, KeysView[str]]:
-    """Build a stats query for calls, automatically using optimized queries when possible.
-
-    This function handles both optimized special-case queries and the general case.
-    Returns a tuple of (query_sql, column_names).
-
-    Args:
-        req: The stats query request
-        param_builder: Parameter builder for query parameterization
-        read_table: Which calls table to read from
-
-    Returns:
-        Tuple of (SQL query string, column names in the result)
-    """
-    aggregated_columns = {"count": "count()"}
-
-    # Try optimized special case queries first
-    if opt_query := _try_optimized_stats_query(req, param_builder, read_table):
-        return (opt_query, aggregated_columns.keys())
-
-    if req.include_total_storage_size:
-        aggregated_columns["total_storage_size_bytes"] = (
-            "sum(coalesce(total_storage_size_bytes, 0))"
-        )
-
-    # For calls_complete, use a flat query (10x+ faster, avoids subquery materialization):
-    #   Fast:  SELECT count() FROM calls_complete WHERE ...
-    #   Slow:  SELECT count() FROM (SELECT id FROM calls_complete WHERE ...)
-    if read_table == ReadTable.CALLS_COMPLETE:
-        query = _build_calls_complete_stats_query(
-            req, param_builder, aggregated_columns
-        )
-        return (query, aggregated_columns.keys())
-
-    # For calls_merged, use subquery wrapping (GROUP BY requires materialization)
-    cq = _build_stats_calls_query(req, read_table)
-    inner_query = cq.as_sql(param_builder)
-    calls_query_sql = f"SELECT {', '.join(aggregated_columns[k] for k in aggregated_columns)} FROM ({inner_query})"
-
-    return (calls_query_sql, aggregated_columns.keys())
-
-
-def _build_stats_calls_query(
-    req: tsi.CallsQueryStatsReq,
-    read_table: ReadTable,
-) -> CallsQuery:
-    """Build a CallsQuery populated with the stats request's filters and conditions."""
-    cq = CallsQuery(
-        project_id=req.project_id,
-        include_total_storage_size=req.include_total_storage_size or False,
-        read_table=read_table,
-    )
-    cq.add_field("id")
-    if req.filter is not None:
-        cq.set_hardcoded_filter(HardCodedFilter(filter=req.filter))
-    if req.query is not None:
-        cq.add_condition(req.query.expr_)
-    if req.limit is not None:
-        cq.set_limit(req.limit)
-    if req.expand_columns is not None:
-        cq.set_expand_columns(req.expand_columns)
-    if req.include_total_storage_size:
-        cq.add_field("total_storage_size_bytes")
-    return cq
-
-
-def _build_calls_complete_stats_query(
-    req: tsi.CallsQueryStatsReq,
-    param_builder: ParamBuilder,
-    aggregated_columns: dict[str, str],
-) -> str:
-    """Build a flat stats query for calls_complete without subquery wrapping.
-
-    Produces SELECT count() FROM calls_complete WHERE ... directly,
-    which is 10x+ faster than the subquery form because ClickHouse evaluates
-    aggregates without materializing intermediate rows.
-
-    Uses CallsQuery._build_query_body for filter/condition/JOIN building,
-    then composes its own aggregate SELECT clause on top.
-    """
-    table_name = get_calls_table_name(ReadTable.CALLS_COMPLETE)
-    cq = _build_stats_calls_query(req, ReadTable.CALLS_COMPLETE)
-
-    # Inject deleted_at sentinel filter for calls_complete (epoch zero = not deleted).
-    # This is needed because _build_query_body doesn't inject it automatically --
-    # that happens in as_sql() which we bypass here.
-    # Use None as the literal so process_operation applies the sentinel value with
-    # proper DateTime64(3) typing (same pattern as as_sql()).
-    cq.add_condition(
-        tsi_query.EqOperation.model_validate(
-            {
-                "$eq": [
-                    {"$getField": "deleted_at"},
-                    {"$literal": None},
-                ]
-            }
-        )
-    )
-
-    # Build the query body (FROM, JOINs, WHERE, etc.) — reuses all existing
-    # filter/condition/join abstractions without CTE optimization overhead
-    body = cq._build_query_body(param_builder, table_name)
-
-    # Build the aggregate SELECT clause
-    stats_parts: list[str] = []
-    for col_name, col_agg in aggregated_columns.items():
-        if col_name == "total_storage_size_bytes":
-            # Inline the total_storage_size expression for the flat query.
-            # Mirrors AggregatedDataSizeField.as_select_sql(use_agg_fn=False),
-            # wrapped in the aggregate function directly.
-            parent_id_field = get_field_by_name("parent_id")
-            parent_null = parent_id_field.null_check_sql(
-                param_builder, table_name, ReadTable.CALLS_COMPLETE, use_agg_fn=False
-            )
-            total_storage_expr = (
-                f"CASE WHEN {parent_null} "
-                f"THEN {ROLLED_UP_CALL_MERGED_STATS_TABLE_NAME}.total_storage_size_bytes "
-                f"ELSE NULL END"
-            )
-            stats_parts.append(f"sum(coalesce({total_storage_expr}, 0)) AS {col_name}")
-        else:
-            stats_parts.append(f"{col_agg} AS {col_name}")
-    stats_select = ", ".join(stats_parts)
-
-    raw_sql = f"""
-    SELECT {stats_select}
-    {body}
-    """
-    return safely_format_sql(raw_sql, logger)
-
-
-def _try_optimized_stats_query(
-    req: tsi.CallsQueryStatsReq,
-    param_builder: ParamBuilder,
-    read_table: ReadTable,
-) -> str | None:
-    """Try to match request to an optimized special-case query.
-
-    Returns optimized query string if a pattern matches, None otherwise.
-    Add new patterns here for common hot-path queries.
-    """
-    # Pattern 1: Simple existence check (limit=1, no filters)
-    if (
-        req.limit == 1
-        and req.filter is None
-        and req.query is None
-        and not req.include_total_storage_size
-    ):
-        return _optimized_project_contains_call_query(
-            req.project_id, param_builder, read_table
-        )
-
-    # Pattern 2: Query with wb_run_id check (limit=1, query present, minimal filter)
-    # Covers common case: checking for runs with wb_run_id not null
-    if (
-        req.limit == 1
-        and req.query is not None
-        and not req.include_total_storage_size
-        and _is_minimal_filter(req.filter)
-    ):
-        return _optimized_wb_run_id_not_null_query(
-            req.project_id, param_builder, read_table
-        )
-
-    return None
-
-
-def _optimized_project_contains_call_query(
-    project_id: str,
-    param_builder: ParamBuilder,
-    read_table: ReadTable,
-) -> str:
-    """Returns a query that checks if the project contains any calls."""
-    table_name = get_calls_table_name(read_table)
-    return safely_format_sql(
-        f"""SELECT
-    toUInt8(count()) AS has_any
-    FROM
-    (
-        SELECT 1
-        FROM {table_name}
-        WHERE project_id = {param_slot(param_builder.add_param(project_id), "String")}
-        LIMIT 1
-    )
-    """,
-        logger,
-    )
-
-
-def _optimized_wb_run_id_not_null_query(
-    project_id: str,
-    param_builder: ParamBuilder,
-    read_table: ReadTable,
-) -> str:
-    """Optimized query for checking existence of calls with wb_run_id not null.
-
-    Uses WHERE clause instead of HAVING to avoid expensive aggregation.
-    """
-    project_id_param = param_builder.add_param(project_id)
-    table_name = get_calls_table_name(read_table)
-    wb_run_id_field = get_field_by_name("wb_run_id")
-    wb_run_id_check = wb_run_id_field.null_check_sql(
-        param_builder, table_name, read_table, use_agg_fn=False, negate=True
-    )
-    deleted_at_field = get_field_by_name("deleted_at")
-    deleted_at_check = deleted_at_field.null_check_sql(
-        param_builder, table_name, read_table, use_agg_fn=False
-    )
-    return f"""
-        SELECT count() FROM (
-            SELECT {table_name}.id AS id
-            FROM {table_name}
-            WHERE {table_name}.project_id = {param_slot(project_id_param, "String")}
-                AND {wb_run_id_check}
-                AND {deleted_at_check}
-            LIMIT 1
-        )
-    """
-
-
-def _is_minimal_filter(filter: tsi.CallsFilter | None) -> bool:
-    """Check if filter has no specific filtering criteria set."""
-    if filter is None:
-        return True
-    return (
-        filter.wb_run_ids is None
-        and filter.wb_user_ids is None
-        and filter.op_names is None
-        and filter.call_ids is None
-        and filter.trace_ids is None
-        and filter.parent_ids is None
-        and filter.trace_roots_only is None
-        and filter.input_refs is None
-        and filter.output_refs is None
-        and filter.thread_ids is None
-        and filter.turn_ids is None
-    )
-
-
-def _format_table_name_with_cluster(table_name: str, cluster_name: str | None) -> str:
-    """Format a table name with ON CLUSTER clause if cluster_name is provided.
-
-    In distributed mode, mutations (UPDATE, DELETE, etc.) must target the local
-    table with the ON CLUSTER clause to execute across all cluster nodes.
-    """
-    if cluster_name:
-        return f"{table_name}{LOCAL_TABLE_SUFFIX} ON CLUSTER {cluster_name}"
-    return table_name
-
-
-def build_calls_complete_update_end_query(
-    table_name: str,
-    project_id_param: str,
-    id_param: str,
-    ended_at_param: str,
-    exception_param: str,
-    output_dump_param: str,
-    summary_dump_param: str,
-    output_refs_param: str,
-    wb_run_step_end_param: str,
-    started_at_param: str | None = None,
-    cluster_name: str | None = None,
-) -> str:
-    """Build the calls_complete UPDATE query for call end data.
-
-    Args:
-        table_name (str): The calls_complete table name.
-        project_id_param (str): Param slot key for project_id.
-        id_param (str): Param slot key for call id.
-        ended_at_param (str): Param slot key for ended_at (Int64 microseconds).
-        exception_param (str): Param slot key for exception.
-        output_dump_param (str): Param slot key for output_dump.
-        summary_dump_param (str): Param slot key for summary_dump.
-        output_refs_param (str): Param slot key for output_refs.
-        wb_run_step_end_param (str): Param slot key for wb_run_step_end.
-        started_at_param (str | None): Optional param slot key for started_at
-            (Int64 microseconds). When provided, enables more efficient queries
-            by utilizing the ClickHouse primary key (project_id, started_at, id).
-        cluster_name (str | None): Optional ClickHouse cluster name for ON CLUSTER
-            clause in distributed mode. When provided, the UPDATE will be executed
-            across all cluster nodes.
-
-    Returns:
-        str: The formatted ClickHouse UPDATE statement.
-
-    Note:
-        started_at and ended_at params are passed as Int64 microseconds since epoch
-        because clickhouse-connect truncates datetime objects to whole seconds.
-        We use fromUnixTimestamp64Micro() to convert back to DateTime64(6).
-    """
-    # Build WHERE clause - include started_at if provided for better primary key usage
-    where_clauses = [f"project_id = {{{project_id_param}:String}}"]
-    if started_at_param is not None:
-        where_clauses.append(
-            f"started_at = fromUnixTimestamp64Micro({{{started_at_param}:Int64}}, 'UTC')"
-        )
-    else:
-        # TODO: try to optimistically parse uuidv7, grabbing timestamps from the ID
-        # then use that to narrow the granules we need to search.
-        pass
-
-    where_clauses.append(f"id = {{{id_param}:String}}")
-    where_clause = " AND ".join(where_clauses)
-
-    # Format table name with ON CLUSTER if cluster_name is provided
-    formatted_table = _format_table_name_with_cluster(table_name, cluster_name)
-
-    # Use fromUnixTimestamp64Micro to convert Int64 microseconds to DateTime64(6)
-    # This preserves full microsecond precision that would be lost with datetime params
-    return f"""
-        UPDATE {formatted_table}
-        SET
-            ended_at = fromUnixTimestamp64Micro({{{ended_at_param}:Int64}}, 'UTC'),
-            exception = {{{exception_param}:String}},
-            output_dump = {{{output_dump_param}:String}},
-            summary_dump = {{{summary_dump_param}:String}},
-            output_refs = {{{output_refs_param}:Array(String)}},
-            wb_run_step_end = {{{wb_run_step_end_param}:UInt64}},
-            updated_at = now64(3)
-        WHERE {where_clause}
-        """
-
-
-def build_calls_complete_delete_query(
-    table_name: str,
-    project_id_param: str,
-    call_ids_param: str,
-    cluster_name: str | None = None,
-) -> str:
-    """Build the calls_complete DELETE query for call end data."""
-    formatted_table = _format_table_name_with_cluster(table_name, cluster_name)
-    raw_sql = f"""
-        DELETE FROM {formatted_table}
-        WHERE project_id = {{{project_id_param}:String}} AND id IN {{{call_ids_param}:Array(String)}}
-        """
-    return safely_format_sql(raw_sql, logger)
-
-
-def build_calls_complete_update_query(
-    table_name: str,
-    project_id_param: str,
-    id_param: str,
-    display_name_param: str,
-    cluster_name: str | None = None,
-) -> str:
-    """Build the calls_complete UPDATE query for call end data."""
-    formatted_table = _format_table_name_with_cluster(table_name, cluster_name)
-    raw_sql = f"""
-        UPDATE {formatted_table}
-        SET display_name = {{{display_name_param}:String}}
-        WHERE project_id = {{{project_id_param}:String}} AND id = {{{id_param}:String}}
-        """
-    return safely_format_sql(raw_sql, logger)
